Implement compose layout modifier for ImeKeyboardBox
